tubers/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_full_version >= '3.7.0'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
cryptography==41.0.3
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.2.4
django-allauth==0.54.0
django-ckeditor==6.7.0
django-filter==23.2
django-js-asset==2.1.0; python_version >= '3.8'
djangocms-admin-style==3.2.4
idna==3.4; python_version >= '3.5'
oauthlib==3.2.2; python_version >= '3.6'
pillow==10.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
pyjwt[crypto]==2.8.0; python_full_version >= '3.7.0'
python3-openid==3.2.0
requests==2.31.0; python_full_version >= '3.7.0'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.7.1; python_version < '3.11'
tzdata==2023.3; sys_platform == 'win32'
urllib3==2.0.4; python_full_version >= '3.7.0'
